Feature/implement realm offline mode

+++ app/src/main/java/com/dega/payconiq/api/ApiService.java
@@ -19,7 +19,7 @@ public class ApiService implements ReposApi {
     }
 
     @Override
-    public Observable<List<Repository>> loadRepositories(int page, int perPage) {
-        return api.loadRepositories(page,perPage);
+    public Observable<List<Repository>> loadRepositories(long page, int perPage) {
+        return api.loadRepositories(page, perPage);
     }
 }
